requirements.txt
Flask
python-dotenv
google-generativeai
PyPDF2
pdfplumber
python-docx
python-pptx
Pillow